+++ components/image/Image.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { computed, ref, watchEffect, nextTick } from 'vue'
+import type { CSSProperties } from 'vue'
 import Space from 'components/space'
 import Spin from 'components/spin'
 import { add, downloadFile } from 'components/utils'
@@ -16,6 +17,7 @@ export interface Props {
   bordered?: boolean // 是否显示边框
   fit?: 'contain' | 'fill' | 'cover' | 'none' | 'scale-down' // 图片在容器内的的适应类型
   preview?: string // 预览文本 string | slot
+  previewImageStyle?: CSSProperties // 自定义预览图片时 img 元素的样式
   spaceProps?: object // Space 组件属性配置，用于配置多张展示图片时的排列方式
   spinProps?: object // Spin 组件属性配置，用于配置图片加载中样式
   zoomRatio?: number // 每次缩放比率
@@ -35,6 +37,7 @@ const props = withDefaults(defineProps<Props>(), {
   bordered: true,
   fit: 'contain',
   preview: '预览',
+  previewImageStyle: () => ({}),
   spaceProps: () => ({}),
   spinProps: () => ({}),
   zoomRatio: 0.1,
@@ -376,7 +379,20 @@ function onSwitchRight() {
               title="下载"
               @click="onDownload"
             >
-              <svg class="icon-svg" focusable="false" data-icon="download" width="1em" height="1em" fill="currentColor" aria-hidden="true" viewBox="64 64 896 896"><path d="M505.7 661a8 8 0 0012.6 0l112-141.7c4.1-5.2.4-12.9-6.3-12.9h-74.1V168c0-4.4-3.6-8-8-8h-60c-4.4 0-8 3.6-8 8v338.3H400c-6.7 0-10.4 7.7-6.3 12.9l112 141.8zM878 626h-60c-4.4 0-8 3.6-8 8v154H214V634c0-4.4-3.6-8-8-8h-60c-4.4 0-8 3.6-8 8v198c0 17.7 14.3 32 32 32h684c17.7 0 32-14.3 32-32V634c0-4.4-3.6-8-8-8z"></path></svg>
+              <svg
+                class="icon-svg"
+                focusable="false"
+                data-icon="download"
+                width="1em"
+                height="1em"
+                fill="currentColor"
+                aria-hidden="true"
+                viewBox="64 64 896 896"
+              >
+                <path
+                  d="M505.7 661a8 8 0 0012.6 0l112-141.7c4.1-5.2.4-12.9-6.3-12.9h-74.1V168c0-4.4-3.6-8-8-8h-60c-4.4 0-8 3.6-8 8v338.3H400c-6.7 0-10.4 7.7-6.3 12.9l112 141.8zM878 626h-60c-4.4 0-8 3.6-8 8v154H214V634c0-4.4-3.6-8-8-8h-60c-4.4 0-8 3.6-8 8v198c0 17.7 14.3 32 32 32h684c17.7 0 32-14.3 32-32V634c0-4.4-3.6-8-8-8z"
+                ></path>
+              </svg>
             </div>
             <div
               class="preview-operation"
@@ -517,7 +533,10 @@ function onSwitchRight() {
           >
             <img
               class="preview-image"
-              :style="`transform: scale3d(${swapX * scale}, ${swapY * scale}, 1) rotate(${rotate}deg);`"
+              :style="[
+                previewImageStyle,
+                `transform: scale3d(${swapX * scale}, ${swapY * scale}, 1) rotate(${rotate}deg);`
+              ]"
               :src="image.src"
               :alt="getImageName(image)"
               @mousedown.prevent="handleMouseDown"