+++ docs/guide/changelog.md
@@ -6,7 +6,7 @@
 
 ## 2.2.6 <Tag color="volcano" size="small">2025-02-19</Tag>
 
-- 优化并更新 [图片 Image](https://themusecatcher.github.io/vue-amazing-ui/guide/components/progress.html) 组件，预览时新增下载图片操作；新增 `disabled` 属性，支持禁用图片预览功能
+- 优化并更新 [图片 Image](https://themusecatcher.github.io/vue-amazing-ui/guide/components/progress.html) 组件，预览时新增下载图片操作；新增 `disabled` `previewImageStyle` 属性，支持：禁用图片预览功能；自定义预览图片时 `img` 元素的样式
 - 优化 [下载文件 downloadFile](https://themusecatcher.github.io/vue-amazing-ui/utils/functions/download-file.html) 工具函数代码
 
 ## 2.2.5 <Tag color="volcano" size="small">2025-02-18</Tag>